# granja-corales-main/SolutionBackEnd/Solution.API/bin/Release/netcoreapp3.1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Solution.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:62718"
    }
  ],
  "paths": {
    "/api/Articulos": {
      "get": {
        "tags": [
          "Articulos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Articulos"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Articulos"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Articulos"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Articulos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Articulos"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Articulos"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Articulos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Categorias": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorias"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorias"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorias"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorias"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Articulos": {
        "type": "object",
        "properties": {
          "idArticulo": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCientifico": {
            "type": "string",
            "nullable": true
          },
          "familia": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "nombreComun": {
            "type": "string",
            "nullable": true
          },
          "dificultad": {
            "type": "string",
            "nullable": true
          },
          "temperamento": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "dieta": {
            "type": "string",
            "nullable": true
          },
          "tamanoMax": {
            "type": "string",
            "nullable": true
          },
          "origen": {
            "type": "string",
            "nullable": true
          },
          "tamanoMinPecera": {
            "type": "string",
            "nullable": true
          },
          "modificadoPor": {
            "type": "string",
            "nullable": true
          },
          "fecModificacion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "imagePath": {
            "type": "string",
            "nullable": true
          },
          "idCategoria": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCategoriaNavigation": {
            "$ref": "#/components/schemas/Categorias"
          }
        },
        "additionalProperties": false
      },
...
      "Categorias": {
        "type": "object",
        "properties": {
          "idCategoria": {
            "type": "integer",
            "format": "int32"
          },
          "tipo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}